style refactor
moving alla images and css files into assets folder and settring the new paths in project's inmports
