+++ src/assets/img/search-arrow.png
