+++ src/assets/stylesheets/Locations/Locations.css
@@ -1,5 +1,5 @@
 .form-container {
-    background-image: url('../../assets/locations.jpg');
+    background-image: url('../../img/locations.jpg');
     background-size: cover;
     background-position: center;
     background-repeat: no-repeat;
