+++ src/assets/stylesheets/RadiusSearch/Object.css
@@ -24,7 +24,7 @@ p {
 .object-item button {
     width: 2rem;
     height: 2rem;
-    background-image: url('../../assets/search-arrow.png');
+    background-image: url('../../img/search-arrow.png');
     border: thin solid grey;
     font-size: medium;
     margin: 0;
