+++ src/assets/stylesheets/RadiusSearch/RadiusSearchForm.css
@@ -55,7 +55,7 @@
 .form button {
     width: 2rem;
     height: 2rem;
-    background-image: url('../../assets/search-arrow.png');
+    background-image: url('../../img/search-arrow.png');
     border: thin solid grey;
     font-size: medium;
     margin-left: 1rem;
